fix: apply some suggestions
